[shrike] Adding test namespaces for all Clojure src namespaces
This commit adds a test namespace for basically any src namespace in
Clojure (i.e. under `src/clj/shrike`). Titan namespaces are not included
in this commit. ClojureScript namespaces are also not included.

Most of the tests aren't yet complete, though some are. In the majority
of cases I've added placeholder tests that currently fail - these will
have to be filled in at a later point in time (i.e. when I'm not on an
airplane with no wifi)